refactor: register workspace dropdown menu to left menu (#90)
- register workspace dropdown menu to left menu from workspace plugin
- rename PUBLIC_WORKSPACE and MANAGEMENT_WORKSPACE

---------
