+++ src/core/public/core_system.ts
@@ -225,7 +225,7 @@ export class CoreSystem {
         targetDomElement: notificationsTargetDomElement,
       });
       const application = await this.application.start({ http, overlays });
-      const workspaces = this.workspaces.start();
+      const workspaces = this.workspaces.start({ application, http });
       const chrome = await this.chrome.start({
         application,
         docLinks,
