+++ src/core/public/workspace/workspaces_service.ts
@@ -2,20 +2,39 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
+
 import { BehaviorSubject } from 'rxjs';
 import { CoreService } from '../../types';
+import { InternalApplicationStart } from '../application';
+import { HttpSetup } from '../http';
 
-/**
- * @public
- */
-export interface WorkspaceStart {
+type WorkspaceMenuRenderFn = ({
+  basePath,
+  getUrlForApp,
+  observables,
+}: {
+  getUrlForApp: InternalApplicationStart['getUrlForApp'];
+  basePath: HttpSetup['basePath'];
+  observables: WorkspaceObservables;
+}) => JSX.Element | null;
+
+export interface WorkspaceObservables {
   currentWorkspaceId$: BehaviorSubject<string>;
   currentWorkspace$: BehaviorSubject<WorkspaceAttribute | null>;
   workspaceList$: BehaviorSubject<WorkspaceAttribute[]>;
   workspaceEnabled$: BehaviorSubject<boolean>;
 }
 
-export type WorkspaceSetup = WorkspaceStart;
+/**
+ * @public
+ */
+export interface WorkspaceSetup extends WorkspaceObservables {
+  registerWorkspaceMenuRender: (render: WorkspaceMenuRenderFn) => void;
+}
+
+export interface WorkspaceStart extends WorkspaceObservables {
+  renderWorkspaceMenu: () => JSX.Element | null;
+}
 
 export interface WorkspaceAttribute {
   id: string;
@@ -32,6 +51,7 @@ export class WorkspaceService implements CoreService<WorkspaceSetup, WorkspaceSt
   private workspaceList$ = new BehaviorSubject<WorkspaceAttribute[]>([]);
   private currentWorkspace$ = new BehaviorSubject<WorkspaceAttribute | null>(null);
   private workspaceEnabled$ = new BehaviorSubject<boolean>(false);
+  private _renderWorkspaceMenu: WorkspaceMenuRenderFn | null = null;
 
   public setup(): WorkspaceSetup {
     return {
@@ -39,15 +59,36 @@ export class WorkspaceService implements CoreService<WorkspaceSetup, WorkspaceSt
       currentWorkspace$: this.currentWorkspace$,
       workspaceList$: this.workspaceList$,
       workspaceEnabled$: this.workspaceEnabled$,
+      registerWorkspaceMenuRender: (render: WorkspaceMenuRenderFn) =>
+        (this._renderWorkspaceMenu = render),
     };
   }
 
-  public start(): WorkspaceStart {
-    return {
+  public start({
+    http,
+    application,
+  }: {
+    application: InternalApplicationStart;
+    http: HttpSetup;
+  }): WorkspaceStart {
+    const observables = {
       currentWorkspaceId$: this.currentWorkspaceId$,
       currentWorkspace$: this.currentWorkspace$,
       workspaceList$: this.workspaceList$,
       workspaceEnabled$: this.workspaceEnabled$,
+    };
+    return {
+      ...observables,
+      renderWorkspaceMenu: () => {
+        if (this._renderWorkspaceMenu) {
+          return this._renderWorkspaceMenu({
+            basePath: http.basePath,
+            getUrlForApp: application.getUrlForApp,
+            observables,
+          });
+        }
+        return null;
+      },
     };
   }
 
@@ -56,5 +97,6 @@ export class WorkspaceService implements CoreService<WorkspaceSetup, WorkspaceSt
     this.currentWorkspaceId$.unsubscribe();
     this.workspaceList$.unsubscribe();
     this.workspaceEnabled$.unsubscribe();
+    this._renderWorkspaceMenu = null;
   }
 }
