+++ src/core/utils/constants.ts
@@ -13,6 +13,6 @@ export enum WorkspacePermissionMode {
   LibraryWrite = 'library_write',
 }
 
-export const PUBLIC_WORKSPACE = 'public';
+export const PUBLIC_WORKSPACE_ID = 'public';
 
-export const MANAGEMENT_WORKSPACE = 'management';
+export const MANAGEMENT_WORKSPACE_ID = 'management';
