+++ src/core/utils/index.ts
@@ -40,6 +40,6 @@ export { DEFAULT_APP_CATEGORIES } from './default_app_categories';
 export {
   WORKSPACE_PATH_PREFIX,
   WorkspacePermissionMode,
-  PUBLIC_WORKSPACE,
-  MANAGEMENT_WORKSPACE,
+  PUBLIC_WORKSPACE_ID,
+  MANAGEMENT_WORKSPACE_ID,
 } from './constants';
