+++ src/plugins/workspace/public/application.tsx
@@ -5,15 +5,13 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { AppMountParameters, CoreStart } from '../../../core/public';
+import { AppMountParameters } from '../../../core/public';
 import { OpenSearchDashboardsContextProvider } from '../../opensearch_dashboards_react/public';
 import { WorkspaceListApp } from './components/workspace_list_app';
 import { WorkspaceCreatorApp } from './components/workspace_creator_app';
 import { WorkspaceUpdaterApp } from './components/workspace_updater_app';
 import { WorkspaceOverviewApp } from './components/workspace_overview_app';
-import { WorkspaceClient } from './workspace_client';
-
-export type Services = CoreStart & { workspaceClient: WorkspaceClient };
+import { Services } from './types';
 
 export const renderListApp = (
   { element, history, appBasePath }: AppMountParameters,
